renamed some test cases. added name prefix check in flow deletion test case

+++ sdk/test/k8s/test_flow.py
@@ -7,14 +7,21 @@ from bettmensch_ai.pipelines import delete_flow, list_flows
 @pytest.mark.delete_flows
 @pytest.mark.order(10)
 def test_delete(test_namespace):
-    """Test the pipeline.delete function"""
+    """Test the delete_flow function"""
 
     flows = list_flows(registered_namespace=test_namespace)
 
     for flow in flows:
-        delete_flow(
-            registered_name=flow.registered_name,
-            registered_namespace=test_namespace,
-        )
+        if flow.registered_name.startswith("test-"):
+            delete_flow(
+                registered_name=flow.registered_name,
+                registered_namespace=test_namespace,
+            )
 
-    assert list_flows(registered_namespace=test_namespace) == []
+    test_flows = [
+        flow
+        for flow in list_flows(registered_namespace=test_namespace)
+        if flow.registered_name.startswith("test-")
+    ]
+
+    assert test_flows == []
